feat: upgrade audio pipeline from 16-bit to 32-bit for FLAC lossless quality
Change swresample output, PCM buffer, PulseAudio (PA_SAMPLE_S32LE),
ALSA (SND_PCM_FORMAT_S32_LE), and visualizer normalization from
16-bit integer to 32-bit integer to preserve full dynamic range
of 24-bit FLAC audio.

+++ include/defs.h
@@ -58,7 +58,7 @@ typedef enum {
 #define UI_DIRTY_LYRICS 0x04
 
 #define CONTROL_COUNT 8
-#define MAX_AUDIO_BUFFER_SIZE (44100 * 2 * sizeof(int16_t))
+#define MAX_AUDIO_BUFFER_SIZE (44100 * 2 * sizeof(int32_t))
 #define MAX_PATH_LEN 512
 #define MAX_TRACKS 1000
 #define MAX_META_LEN 256
@@ -296,7 +296,7 @@ void process_pending_ui_refresh(void);
 void reap_finished_playback_thread(void);
 void process_pending_playback_action(void);
 void reset_visualizer_state(void);
-void push_visualizer_samples(const int16_t *samples, int frame_count, int channels);
+void push_visualizer_samples(const int32_t *samples, int frame_count, int channels);
 void get_visualizer_snapshot(int *levels, int *peaks, int max_levels, uint64_t *last_update_ms);
 
 void apply_color_theme(void);
